Feature/cart

+++ components/cart/CartItems.tsx
@@ -7,6 +7,8 @@ import { CartService } from "@/services/api/cart.service";
 import CreateAxiosInstanceWithLoader from "@/services/utility/axios-with-loader.service";
 import { useRef, useState } from "react";
 import debounce from 'lodash/debounce';
+import { useUserDataProvider } from "@/providers/UserDataProvider";
+import { useTranslation } from "react-i18next";
 
 interface CartItemProps {
   item: any;
@@ -17,7 +19,9 @@ export function CartItems({ item, showSubtotal = true }: CartItemProps) {
   const { updateQuantity, removeFromCart, updateCartCredentials } = useCart();
   const axiosInstanceWithoutLoader = CreateAxiosInstanceWithLoader(false, false);
   const [isCartItemUpdating, setIsCartItemUpdating] = useState<boolean>(false);
-  
+  const { isAuthenticated } = useUserDataProvider();
+  const { t } = useTranslation("common");
+
   const useDebouncedUpdateCartItemQuantity = () => {
     const debouncedFn = useRef(
       debounce(async (cartKey: string, itemId: number, quantity: number, updateQuantityFn: any, setLoadingFn: any) => {
@@ -25,7 +29,7 @@ export function CartItems({ item, showSubtotal = true }: CartItemProps) {
           setLoadingFn(true);
           const response = await CartService.UpdateCartItem(itemId, quantity, axiosInstanceWithoutLoader);
           updateCartCredentials(response.cartToken, response.nonce);
-          
+
           const item = response.data.items.find((x: any) => x.id === itemId);
           updateQuantityFn(itemId, Math.max(1, item?.quantity));
           setLoadingFn(false);
@@ -35,21 +39,30 @@ export function CartItems({ item, showSubtotal = true }: CartItemProps) {
         }
       }, 500) // 500ms debounce time
     ).current;
-  
+
     return debouncedFn;
   };
-  
+
   const debouncedUpdate = useDebouncedUpdateCartItemQuantity();
-  const handleQuantityChange = (cartKey: string, itemId: number, newQuantity: number) => {
+  const handleQuantityChange = (cartKey: string, itemId: number, newQuantity: number, bundleId: number) => {
     if (newQuantity < 1) {
       return;
     }
 
-    updateQuantity(itemId, Math.max(1, newQuantity));
+    if (!isAuthenticated) {
+      updateQuantity(itemId, Math.max(1, newQuantity), bundleId);
+      return;
+    }
+
+    updateQuantity(itemId, Math.max(1, newQuantity), bundleId);
     debouncedUpdate(cartKey, itemId, newQuantity, updateQuantity, setIsCartItemUpdating);
   };
 
   const handleCartItemRemove = async (cartKey: string, itemId: number) => {
+    if (!isAuthenticated) {
+      removeFromCart(itemId);
+      return;
+    }
     try {
       const response = await CartService.DeleteCartItem(itemId);
       updateCartCredentials(response.cartToken, response.nonce);
@@ -59,11 +72,18 @@ export function CartItems({ item, showSubtotal = true }: CartItemProps) {
     }
   };
 
+  // const showControls = () => {
+  //   // Show controls (remove and quantity) if:
+  //   // 1. Item has no bundleId (regular item) OR
+  //   // 2. Item's id matches its bundleId (main bundle item)
+  //   return !item.bundleId || item.productId === item.bundleId;
+  // };
+
   return (
     <div className="flex gap-4 py-4">
       <div className="relative w-20 h-20">
         <Image
-          src={item.image?.src || item?.image ||  "/placeholder.svg"}
+          src={item.image?.src || item?.image || "/placeholder.svg"}
           alt={item.name}
           height={80}
           width={80}
@@ -83,22 +103,24 @@ export function CartItems({ item, showSubtotal = true }: CartItemProps) {
         <div className="mt-2 flex items-center gap-4">
           <div className="flex items-center border rounded">
             <button
-              onClick={() => handleQuantityChange(item.cartKey, item.id, item.quantity - 1)}
+              onClick={() => handleQuantityChange(item.cartKey, item.id, item.quantity - 1, item.bundleId)}
               className="px-2 py-1 hover:bg-muted"
+
             >
               -
             </button>
             <span className="w-8 text-center">{item.quantity}</span>
             <button
 
-              onClick={() => handleQuantityChange(item.cartKey, item.id, item.quantity + 1)}
+              onClick={() => handleQuantityChange(item.cartKey, item.id, item.quantity + 1, item.bundleId)}
               className="px-2 py-1 hover:bg-muted"
+
             >
               +
             </button>
           </div>
           <div className="text-primary font-medium">
-            {typeof item.price === "number" ? item.price.toFixed(3) : ""} KD
+            {item.quantity} x {typeof item.price === "number" ? item.price.toFixed(3) : ""} KD
           </div>
           {showSubtotal && (
             <div className="ml-auto text-primary font-medium">
